printer_d450: Tweaks after testing on hardware
Image data commands need a separate function - it's not "0x1b 0x16"
preceding the data, it's just 0x16 (woops!)

The bits in an image data byte need to be reversed.

Something isn't quite right with alignment - presumably something to do
with "top-of-form" detection/flagging. More experimentation needed to
figure this out.