Add printer fudge factor
The printer seems to have an unavoidable 5mm margin on the right edge,
which I can't fix: It doesn't back up far enough from "top-of-form",
and that's hard-coded in the printer (according to the TRM).

So, just subtract 10mm from the total label width. I could plumb this in
more neatly by adding a "margin" setting on the Label class.