WIP: Try a different defer mechanism, waiting for inactivity
Now instead of a fixed delay, this waits for nothing to happen for
DELAY_TIME, and then renders.

This should "feel" better I think.

However, all of this deferring could lead to issues where a stale image
gets printed; so that needs to be avoided somehow (here, I disable the
print button, but that doesn't work for the initial render of the
NameBadge).

I probably need "synchronous" and "asynchronous" paths (e.g. the buttons
on TroveLabel should be synchronous), and link print buttons to the
"validity" of the preview.

Sounds messy!